pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "learning app for students"
authors = ["cyril-pierro <fiopapa32@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.121.2"
pydantic-settings = "2.0.3"
authlib = "^1.6.5"
redis = "^7.0.1"
fastapi-mail = "1.4.1"
sqlalchemy = "^2.0.44"
python-jose = "^3.5.0"
google-api-python-client = "^2.115.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
